tools/deploy/requirements.txt
PyYAML==6.0
rich==13.4.2
typer==0.9.0
typing_extensions==4.7.1